Isolate test-only utility.
